requirements.txt
gymnasium[mujoco]
shimmy
numba
shapely
numpy==2.1.3
gym
pyglet
PyYAML
stable-baselines3[extra]
matplotlib
stable-baselines3
optuna
pytest
torchrl
rich
scipy
tensorboard

